Register plugins more explicitly in registry rather than relying on init() blocks at import time
