+++ Justfile
@@ -14,8 +14,7 @@ tidy:
   @cd pkg/domain && go mod tidy
 
 compile-plugins:
-  ./tools/compile-plugins/compile-plugin.sh list.go internal/connectors/plugins/public public github.com/formancehq/payments/internal/connectors/plugins/public
-  ./tools/compile-plugins/compile-plugin.sh list.go ee/plugins plugins github.com/formancehq/payments/ee/plugins
+  ./tools/compile-plugins/compile-plugin.sh
 
 [group('openapi')]
 validate-openapi:
