--- ee.go
@@ -1,5 +1,0 @@
-//go:build ee
-
-package main
-
-import _ "github.com/formancehq/payments/ee/plugins"
