--- ee/plugins/list.go
@@ -1,9 +1,0 @@
-package plugins
-
-import (
-    _ "github.com/formancehq/payments/ee/plugins/bankingbridge"
-    _ "github.com/formancehq/payments/ee/plugins/bitstamp"
-    _ "github.com/formancehq/payments/ee/plugins/coinbaseprime"
-    _ "github.com/formancehq/payments/ee/plugins/fireblocks"
-    _ "github.com/formancehq/payments/ee/plugins/routable"
-)
