+++ internal/connectors/plugins/errors.go
@@ -1,15 +1,13 @@
 package plugins
 
-import (
-	"errors"
-)
+import pkgplugins "github.com/formancehq/payments/pkg/domain/plugins"
 
 var (
-	ErrNotImplemented       = errors.New("not implemented")
-	ErrNotYetInstalled      = errors.New("not yet installed")
-	ErrInvalidClientRequest = errors.New("invalid client request")
-	ErrUpstreamRatelimit    = errors.New("rate limited by upstream server")
-	ErrUpstreamTimeout      = errors.New("upstream timeout")
-	ErrUpstreamRetryAfter   = errors.New("upstream asked to retry later")
-	ErrCurrencyNotSupported = errors.New("currency not supported")
+	ErrNotImplemented       = pkgplugins.ErrNotImplemented
+	ErrNotYetInstalled      = pkgplugins.ErrNotYetInstalled
+	ErrInvalidClientRequest = pkgplugins.ErrInvalidClientRequest
+	ErrUpstreamRatelimit    = pkgplugins.ErrUpstreamRatelimit
+	ErrUpstreamTimeout      = pkgplugins.ErrUpstreamTimeout
+	ErrUpstreamRetryAfter   = pkgplugins.ErrUpstreamRetryAfter
+	ErrCurrencyNotSupported = pkgplugins.ErrCurrencyNotSupported
 )
