+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"github.com/formancehq/payments/cmd"
-	_ "github.com/formancehq/payments/internal/connectors/plugins/public"
 )
 
 func main() {
