+++ pkg/domain/plugins/plugin.go
@@ -1,0 +1,8 @@
+package plugins
+
+import "github.com/formancehq/payments/pkg/domain/models"
+
+type Plugin struct {
+	models.PSPPlugin
+	models.OpenBankingPlugin
+}
